+++ src/components/schedule-tables/__private__/index.ts
@@ -1,0 +1,1 @@
+export * from "./ScheduleTable";
